feat(port): move historian into src/contextledger/, relocate templates & examples, add ProjectsV2 stub
