+++ .github/workflows/weekly-history.yml
@@ -46,7 +46,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          python -m pip install -e .
 
       - name: Determine metadata
         id: meta
@@ -54,13 +54,12 @@ jobs:
 
       - name: Generate history
         env:
-          PYTHONPATH: ${{ github.workspace }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SINCE_INPUT: ${{ inputs.since }}
           TEMPLATE_INPUT: ${{ inputs.template }}
         run: |
           SINCE_VALUE="${SINCE_INPUT:-7d}"
-          BASE_CMD=(python -m scripts.generate_history --since "$SINCE_VALUE" --until now --output docs/history --debug-scan)
+          BASE_CMD=(python -m contextledger.cli --since "$SINCE_VALUE" --until now --output docs/history)
           if [ -n "$TEMPLATE_INPUT" ]; then
             BASE_CMD+=(--template "$TEMPLATE_INPUT")
           fi
@@ -69,7 +68,7 @@ jobs:
       - name: Check for changes
         id: changes
         run: |
-          if git diff --quiet -- docs/history docs/context/context-index.json; then
+          if git diff --quiet -- docs/history; then
             echo "changed=false" >> "$GITHUB_OUTPUT"
           else
             echo "changed=true" >> "$GITHUB_OUTPUT"
@@ -89,10 +88,9 @@ jobs:
             Automated history snapshot generated by the Git Historian.
 
             - Since: ${{ inputs.since || '7d' }}
-            - Template: ${{ inputs.template || 'docs/history/HISTORY_TEMPLATE.md' }}
+            - Template: ${{ inputs.template || 'templates/weekly.md' }}
           add-paths: |
             docs/history
-            docs/context/context-index.json
 
       - name: No changes detected
         if: steps.changes.outputs.changed != 'true'
