+++ docs/git-historian.md
@@ -7,7 +7,7 @@ output for your team.
 ## Prerequisites
 
 * Python 3.10+
-* `requests` (installed automatically when you run `pip install -r requirements.txt`)
+* `requests` (installed automatically when you run `pip install -e .`)
 * A GitHub token with `repo` scope when running outside GitHub Actions. Export it as `GITHUB_TOKEN`.
 * Optional integrations:
   * Set `HISTORIAN_ENABLE_JIRA=true` with relevant Jira credentials for linking stories to commits.
@@ -17,22 +17,21 @@ output for your team.
 
 ```bash
 # Activate your virtual environment and install dependencies if needed
-pip install -r requirements.txt
+python -m pip install -e .
 
 # Generate a history snapshot from the last 7 days
 export GITHUB_TOKEN=<your-token>
-export PYTHONPATH=$(pwd)
-python -m scripts.generate_history --since 7d --until now --output docs/history --debug-scan
+python -m contextledger.cli --since 7d --until now --output docs/history --debug-scan
 ```
 
-* The script creates `docs/history/YYYY-MM-DD-checkin.md` using [`docs/history/HISTORY_TEMPLATE.md`](history/HISTORY_TEMPLATE.md).
+* The script creates `docs/history/YYYY-MM-DD-checkin.md` using [`templates/weekly.md`](../templates/weekly.md).
 * Use `--debug-scan` to automatically enable DEBUG logging and emit additional scan diagnostics (counts, resolved paths).
   Pass `--log-level` if you need a different verbosity.
 * Use `--since` with ISO timestamps (`2025-01-01T00:00:00Z`) or relative windows (`14d`, `48h`).
 * Use `--until now` (default) or a specific ISO timestamp (`2025-01-15`) to cap the window end.
 * Use `--repo owner/name` to override automatic repository detection.
 * Use `--config <path>` to load a different historian configuration (defaults to `config/defaults.yml`).
-* **Tip:** If you see `ModuleNotFoundError: No module named 'scripts'`, confirm you are running from the repository root and that `PYTHONPATH` includes the root (e.g., `export PYTHONPATH=$(pwd)`).
+* **Tip:** If you see `ModuleNotFoundError: No module named 'contextledger'`, ensure the package is installed in your environment (`python -m pip install -e .`).
 
 ### Collector overview
 
@@ -57,7 +56,7 @@ and can also be triggered manually (`workflow_dispatch`). It performs the follow
 2. Lint the GitHub workflow definitions with [`reviewdog/action-actionlint`](https://github.com/reviewdog/action-actionlint)
    pinned to commit `93dc1f9bc10856298b6cc1a3b3239cfbbb87fe4b` (release `v1.67.0`) and `fail_level: error`
    so any detected issues fail fast.
-3. Run `PYTHONPATH=$(pwd) python -m scripts.generate_history --since 7d --until now --output docs/history --debug-scan`.
+3. Run `python -m contextledger.cli --since 7d --until now --output docs/history`.
 4. Commit changes in `docs/history/*.md` on a branch named `auto/history-<date>`.
 5. Open a pull request summarizing the update.
 
@@ -104,4 +103,4 @@ To run the workflow manually:
 
 * Link the generated check-ins from your Release Copilot dashboards or onboarding docs.
 * Extend the generator to publish artifacts to S3 and enable the S3 collector in `config/defaults.yml`.
-* Use the machine-readable index at `docs/context/context-index.json` (generated by the script) to support RAG pipelines.
+* Use the machine-readable index at `docs/history/context-index.json` (generated by the script) to support RAG pipelines.
