--- docs/history/README.md
@@ -1,11 +1,0 @@
-# Git Historian Snapshots
-
-The `docs/history` folder contains weekly and on-demand project history check-ins generated by the Git Historian tooling.
-
-Each file follows the `YYYY-MM-DD-checkin.md` naming convention and is produced from [`scripts/generate_history.py`](../../scripts/generate_history.py) using the [history template](HISTORY_TEMPLATE.md). The generator aggregates completed work (merged pull requests and closed issues), in-progress and backlog items from GitHub Projects, decisions captured in issue/PR comments, and build artifacts (from GitHub Actions and optional S3 buckets).
-
-In addition to the human-readable Markdown snapshots, the generator keeps an updated machine-readable index at
-[`docs/context/context-index.json`](../context/context-index.json) so other automation (such as RAG pipelines) can easily discover the
-latest check-ins.
-
-See [Git Historian (How to Run)](../git-historian.md) for full usage instructions, including how to run the generator locally, schedule automated updates, and customize the output.
