+++ pyproject.toml
@@ -10,6 +10,13 @@ readme = "README.md"
 requires-python = ">=3.9"
 license = {text = "Apache-2.0"}
 authors = [{ name = "Shayne Vandersloot" }]
+dependencies = [
+    "requests",
+    "PyYAML",
+]
+
+[project.optional-dependencies]
+s3 = ["boto3"]
 
 [project.urls]
 Homepage = "https://github.com/dutchsloot84/ContextLedger"
