+++ src/contextledger/github/__init__.py
@@ -1,0 +1,5 @@
+"""GitHub integration helpers for ContextLedger."""
+
+from .projects import ProjectsV2Client
+
+__all__ = ["ProjectsV2Client"]
